Agents/TaskManagement-agent/requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# LLM Integration
google-generativeai==0.8.5

# Database
asyncpg==0.29.0

# HTTP Client
httpx==0.28.1
requests==2.32.3

# Environment
python-dotenv==1.0.1

